Removing unnecessary imports, trimming dependencies and adding lower bounds
