+++ amazonka-s3-encryption/amazonka-s3-encryption.cabal
@@ -27,6 +27,7 @@ library
 
     exposed-modules:
           Network.AWS.S3.Encryption
+        , Network.AWS.S3.Encryption.Body
         , Network.AWS.S3.Encryption.Decrypt
         , Network.AWS.S3.Encryption.Encrypt
         , Network.AWS.S3.Encryption.Envelope
@@ -34,23 +35,18 @@ library
         , Network.AWS.S3.Encryption.Types
 
     build-depends:
-          amazonka      == 1.3.0.*
-        , amazonka-core == 1.3.0.*
-        , amazonka-kms  == 1.3.0.*
-        , amazonka-s3   == 1.3.0.*
-        , aeson
-        , base          >= 4.7     && < 5
-        , cryptonite
-        , bytestring
-        , conduit-combinators
-        , mono-traversable
-        , http-client
-        , unordered-containers
-        , memory
-        , case-insensitive
-        , lens
-        , text
-        , mtl
-        , conduit
-        , exceptions
-        , resourcet
+          aeson
+        , amazonka             == 1.3.0.*
+        , amazonka-core        == 1.3.0.*
+        , amazonka-kms         == 1.3.0.*
+        , amazonka-s3          == 1.3.0.*
+        , base                 >= 4.7     && < 5
+        , bytestring           >= 0.9
+        , case-insensitive     >= 1.2
+        , conduit-combinators  >= 0.3
+        , cryptonite           >= 0.4
+        , exceptions           >= 0.6
+        , lens                 >= 4.4
+        , memory               >= 0.6
+        , text                 >= 1.1
+        , unordered-containers >= 0.2.5
