+++ amazonka-s3-encryption/src/Network/AWS/S3/Encryption/Encrypt.hs
@@ -1,12 +1,11 @@
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE RankNTypes        #-}
-{-# LANGUAGE RecordWildCards   #-}
 {-# LANGUAGE TupleSections     #-}
 {-# LANGUAGE TypeFamilies      #-}
 
 -- |
--- Module      : Network.AWS.S3.Encrypt
+-- Module      : Network.AWS.S3.Encryption.Encrypt
 -- Copyright   : (c) 2013-2015 Brendan Hay
 -- License     : Mozilla Public License, v. 2.0.
 -- Maintainer  : Brendan Hay <brendan.g.hay@gmail.com>
@@ -15,23 +14,15 @@
 --
 module Network.AWS.S3.Encryption.Encrypt where
 
-import           Control.Arrow
-import           Control.Monad.Catch
-import           Control.Monad.Reader
+import           Control.Monad
 import           Control.Monad.Trans.AWS
-import           Control.Monad.Trans.Resource
-import qualified Data.Aeson.Types                       as Aeson
 import           Data.Coerce
-import           Data.Conduit
 import           Data.Proxy
 import           Network.AWS.Prelude                    hiding (coerce)
-import           Network.AWS.Response
 import           Network.AWS.S3
-import qualified Network.AWS.S3                         as S3
 import           Network.AWS.S3.Encryption.Envelope
 import           Network.AWS.S3.Encryption.Instructions
 import           Network.AWS.S3.Encryption.Types
-import           System.IO
 
 -- | Note about how it doesn't attach metadata by default.
 -- You can re-set the location and then discard the PutInstructions request.
